Fix warnings.
A lot of formatting string warnings fixed, especially ones related to
the NS[U]Integer, which isn't allowed in format strings.

Deprecation warnings squelched in Athena.  All of Carbon, including the
resource manager, is now deprecated, so we otherwise get a bunch of
irrelevant warnings from that.

All vorbis warnings squelched; we wouldn't fix those here.

A couple of others too.

+++ Archivers/Lua/LuaArchiver.m
@@ -115,7 +115,7 @@
         NSInteger idx = (isZeroIndexed?0:1);
         for (id<LuaCoding> obj in array) {
             [self encodeObject:obj
-                        forKey:[NSString stringWithFormat:@"[%d]", idx]];
+                        forKey:[NSString stringWithFormat:@"[%ld]", (long)idx]];
             idx++;
         }
         [self indent];
@@ -203,7 +203,7 @@
 
 - (void) encodeInteger:(NSInteger)value {
     @synchronized(self) {
-        [data appendFormat:@"%lli", value];
+        [data appendFormat:@"%li", (long)value];
     }
 }
 
@@ -211,7 +211,7 @@
     @synchronized(self) {
         [keyStack addObject:key];
         [self indent];
-        [data appendFormat:@"%@ = %lli;\n", key, value];
+        [data appendFormat:@"%@ = %li;\n", key, (long)value];
         [keyStack removeLastObject];
     }
 }
